--- Backend/.vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "jira-plugin.workingProject": ""
-}
