--- Backend/fileUpload/bin/python3
@@ -1,1 +1,0 @@
-/usr/bin/python3
