+++ loginapp/templates/base.html
@@ -39,6 +39,9 @@
                                 Logout
                             </a>
                             {% endif %}
+                            <a href="/uploadme" class="navbar-item">
+                                Upload Files
+                            </a>
                         </div>
                     </div>
                 </div>
